requirements.txt
pandas
openpyxl
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
mistralai
rapidfuzz
